xtask: split transition.rs into transition/ submodules
Pure code move: mod.rs (ParsedRow/ParsedScoreboard/FingerprintFormat),
fingerprint.rs, parse.rs, diff.rs, render_markdown.rs, render_text.rs.
